power_spectrum correto

+++ ref_model/c/lib/mel.h
@@ -18,7 +18,7 @@ void apply_filterbank(
 
 void optimization_filterbank_q15(int32_t filterbank[NUM_FILTERS][OPTIMIZATION_SIZE]);
 void optimization_apply_q15(
-    int32_t power_spectrum_frame[NFFT/2 + 1], 
+    int64_t power_spectrum_frame[NFFT/2 + 1], 
     int32_t energies_q15[NUM_FILTERS]
 );
